refactor(credits): replace direct credit price retrieval with agentService method and remove unused credit service files
